Refactor SingleActivity, implement is with sessions for activity
