+++ frog/imports/client/SingleActivity/style.js
@@ -1,0 +1,53 @@
+const style = {
+  card: {
+    minWidth: 275,
+    maxWidth: 900,
+    margin: 'auto',
+    marginTop: 16,
+    marginBottom: 16,
+    padding: 16
+  },
+  welcome_card: {
+    minWidth: 275,
+    maxWidth: 900,
+    minHeight: 300,
+    margin: 'auto',
+    marginTop: 16,
+    marginBottom: 16,
+    padding: 16,
+    backgroundImage: "url('/table_bg.png')",
+    backgroundPosition: 'right bottom',
+    backgroundRepeat: 'no-repeat'
+  },
+  icon: {
+    height: 96,
+    width: 96,
+    left: '50%',
+    transform: 'translateX(-50%)'
+  },
+  shortDesc: {
+    position: 'inherit'
+  },
+  button: {
+    right: 0
+  },
+  tile: {
+    background: 'rgba(0, 0, 0, 0.3)',
+    backgroundClip: 'content-box'
+  },
+  navbar: {
+    flexDirection: 'row-reverse'
+  },
+  logo: {
+    position: 'absolute',
+    left: 16
+  },
+  padded_text: {
+    margin: 8
+  },
+  back_button: {
+    margin: 8
+  }
+};
+
+export { style };
